+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 1. [15777](https://github.com/influxdata/influxdb/pull/15777): Fix long startup when running 'influx help'
 1. [15713](https://github.com/influxdata/influxdb/pull/15713): Mock missing Flux dependencies when creating tasks
 1. [15731](https://github.com/influxdata/influxdb/pull/15731): Ensure array cursor iterator stats accumulate all cursor stats
+1. [15866](https://github.com/influxdata/influxdb/pull/15866): Do not show Members section in Cloud environments
 
 ### UI Improvements
 1. [15809](https://github.com/influxdata/influxdb/pull/15809): Redesign cards and animations on getting started page